Fixing Total number for Signature
The signature is asking for the total number with two decimal precision and the '.' as the decimal separator. However, when the number reaches the 1000 value, the signature shouldn't contain the regional separator ','. (valid: 1000.32 | invalid: 1,000.32).
According to the powershell documentation about -f function:
:f5 = fix to 5 places
:n	Number (:nP precision=number of decimal places), includes culture separator for thousands 1,000.00

BTW, Thank you very much for sharing this code!